refactor: Merge `markdown_katex.js` test file with `markdown.js`.
We split them in the commit: c50dbf82974cafc4eeb00c5d1fe86ab18e000374
because at that time we thought markdown_katex was gonna
require rewireproxy magic in the later commits.
But we later removed the rewiremock dependency in the
commit: 30c7108955d86d9b860ef6e746ef1e707dbfe62f.

So now, we can safely merge `markdown_katex.js` test file
with `markdown.js` test file.
